Refactor rewriter
