+++ lib/mry.rb
@@ -2,17 +2,7 @@ require 'yaml_rewriter'
 require 'optparse'
 
 require "mry/version"
-require 'mry/rewriter_master'
-require 'mry/rewriter_0_50_0'
-require 'mry/rewriter_0_49_0'
-require 'mry/rewriter_0_47_0'
-require 'mry/rewriter_0_46_0'
-require 'mry/rewriter_0_45_0'
-require 'mry/rewriter_0_44_0'
-require 'mry/rewriter_0_43_0'
-require 'mry/rewriter_0_42_0'
-require 'mry/rewriter_0_41_0'
-require 'mry/rewriter_0'
+require 'mry/rewriters'
 require 'mry/runner'
 require 'mry/cli'
 require 'mry/added_cops'
