--- lib/mry/rewriter_0_41_0.rb
@@ -1,5 +1,0 @@
-module Mry
-  class Rewriter_0_41_0 < YAMLRewriter::Rewriter
-    define_rule ['Style/DeprecatedHashMethods' => 'Style/PreferredHashMethods']
-  end
-end
