--- lib/mry/rewriter_0_42_0.rb
@@ -1,5 +1,0 @@
-module Mry
-  class Rewriter_0_42_0 < YAMLRewriter::Rewriter
-    # 0.42.0 release does not have renaming.
-  end
-end
