--- lib/mry/rewriter_0_45_0.rb
@@ -1,5 +1,0 @@
-module Mry
-  class Rewriter_0_45_0 < YAMLRewriter::Rewriter
-    # 0.45.0 release does not have renaming.
-  end
-end
