--- lib/mry/rewriter_0_46_0.rb
@@ -1,5 +1,0 @@
-module Mry
-  class Rewriter_0_46_0 < YAMLRewriter::Rewriter
-    define_rule ['Performance/SortWithBlock' => 'Performance/CompareWithBlock']
-  end
-end
