--- lib/mry/rewriter_0_49_0.rb
@@ -1,80 +1,0 @@
-module Mry
-  class Rewriter_0_49_0 < YAMLRewriter::Rewriter
-    layout_cops = %w[
-      Layout/AccessModifierIndentation
-      Layout/AlignArray
-      Layout/AlignHash
-      Layout/AlignParameters
-      Layout/BlockEndNewline
-      Layout/CaseIndentation
-      Layout/ClosingParenthesisIndentation
-      Layout/CommentIndentation
-      Layout/DotPosition
-      Layout/ElseAlignment
-      Layout/EmptyLineAfterMagicComment
-      Layout/EmptyLineBetweenDefs
-      Layout/EmptyLinesAroundAccessModifier
-      Layout/EmptyLinesAroundBeginBody
-      Layout/EmptyLinesAroundBlockBody
-      Layout/EmptyLinesAroundClassBody
-      Layout/EmptyLinesAroundExceptionHandlingKeywords
-      Layout/EmptyLinesAroundMethodBody
-      Layout/EmptyLinesAroundModuleBody
-      Layout/EmptyLines
-      Layout/EndOfLine
-      Layout/ExtraSpacing
-      Layout/FirstArrayElementLineBreak
-      Layout/FirstHashElementLineBreak
-      Layout/FirstMethodArgumentLineBreak
-      Layout/FirstMethodParameterLineBreak
-      Layout/FirstParameterIndentation
-      Layout/IndentArray
-      Layout/IndentAssignment
-      Layout/IndentationConsistency
-      Layout/IndentationWidth
-      Layout/IndentHash
-      Layout/IndentHeredoc
-      Layout/InitialIndentation
-      Layout/LeadingCommentSpace
-      Layout/MultilineArrayBraceLayout
-      Layout/MultilineAssignmentLayout
-      Layout/MultilineBlockLayout
-      Layout/MultilineHashBraceLayout
-      Layout/MultilineMethodCallBraceLayout
-      Layout/MultilineMethodCallIndentation
-      Layout/MultilineMethodDefinitionBraceLayout
-      Layout/MultilineOperationIndentation
-      Layout/RescueEnsureAlignment
-      Layout/SpaceAfterColon
-      Layout/SpaceAfterComma
-      Layout/SpaceAfterMethodName
-      Layout/SpaceAfterNot
-      Layout/SpaceAfterSemicolon
-      Layout/SpaceAroundBlockParameters
-      Layout/SpaceAroundEqualsInParameterDefault
-      Layout/SpaceAroundKeyword
-      Layout/SpaceAroundOperators
-      Layout/SpaceBeforeBlockBraces
-      Layout/SpaceBeforeComma
-      Layout/SpaceBeforeComment
-      Layout/SpaceBeforeFirstArg
-      Layout/SpaceBeforeSemicolon
-      Layout/SpaceInLambdaLiteral
-      Layout/SpaceInsideArrayPercentLiteral
-      Layout/SpaceInsideBlockBraces
-      Layout/SpaceInsideBrackets
-      Layout/SpaceInsideHashLiteralBraces
-      Layout/SpaceInsideParens
-      Layout/SpaceInsidePercentLiteralDelimiters
-      Layout/SpaceInsideRangeLiteral
-      Layout/SpaceInsideStringInterpolation
-      Layout/Tab
-      Layout/TrailingBlankLines
-      Layout/TrailingWhitespace
-    ]
-
-    layout_cops.each do |cop|
-      define_rule [cop.sub('Layout', 'Style') => cop]
-    end
-  end
-end
