--- lib/mry/rewriter_master.rb
@@ -1,4 +1,0 @@
-module Mry
-  class Rewriter_Master < YAMLRewriter::Rewriter
-  end
-end
